requirements.txt
pypco>=1.2.0,<2.0
python-dotenv>=1.0.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-httplib2>=0.2.0
httpx>=0.27.0
openpyxl>=3.1.0
PyYAML>=6.0.0
